third_party/blink/web_tests/platform/generic/fast/forms/date/date-stepup-stepdown-expected.txt

Check stepUp() and stepDown() behavior for type=date.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Invalid value
PASS stepUp("", null, null) is "1970-01-02"
PASS stepDown("", null, null) is "1969-12-31"
Non-number arguments
PASS stepUp("2010-02-10", null, null, "0") is "2010-02-10"
PASS stepDown("2010-02-10", null, null, "0") is "2010-02-10"
PASS stepUp("2010-02-10", null, null, "foo") is "2010-02-10"
PASS stepDown("2010-02-10", null, null, "foo") is "2010-02-10"
PASS stepUp("2010-02-10", null, null, null) is "2010-02-10"
PASS stepDown("2010-02-10", null, null, null) is "2010-02-10"
Normal cases
PASS stepDown("1970-01-04", "2", null) is "1970-01-03"
PASS stepUp("2010-02-10", null, null) is "2010-02-11"
PASS stepDown("2010-02-10", null, null) is "2010-02-09"
PASS stepUp("2010-02-10", null, null, 10) is "2010-02-20"
PASS stepDown("2010-02-10", null, null, 11) is "2010-01-30"
PASS stepUp("1970-01-01", "4", null, 2) is "1970-01-09"
PASS stepDown("1970-01-01", "4", null, 3) is "1969-12-20"
Step=any
PASS stepUp("2010-02-10", "any", null) threw exception InvalidStateError: Failed to execute 'stepUp' on 'HTMLInputElement': This form element does not have an allowed value step..
PASS stepDown("2010-02-10", "any", null) threw exception InvalidStateError: Failed to execute 'stepDown' on 'HTMLInputElement': This form element does not have an allowed value step..
Overflow/underflow
PASS stepUp("2010-02-10", "3.40282346e+38", null) is "1970-01-01"
PASS stepDown("2010-02-10", "3.40282346e+38", null) is "1970-01-01"
PASS stepUp("2010-02-10", "1", "2010-02-10") is "2010-02-10"
PASS stepDown("2010-02-10", "1", "2010-02-10") is "2010-02-10"


PASS successfullyParsed is true

TEST COMPLETE
